requirements.txt
fastapi==0.110.0
python-dotenv==1.0.1
selenium==4.17.2
uvicorn==0.28.0
webdriver-manager==4.0.1
